Bring the branch up to date with master.

- AzureObjectStorage: master made `connection_params` a versioned holder,
  so the branch's delete and copy code read the container through
  `connection_params.get()`.
- ReadBufferFromAzureBlobStorage.cpp: keep both new error codes and
  master's returned-range check.
- copyS3File.cpp: keep master's `idempotency_id` and the branch's
  `copy_settings` members.
- SettingsChangesHistory.cpp: take master's 26.10 block and add the
  branch's `azure_use_native_copy` row; the external DISTINCT settings stay
  in 26.9 as on master.

+++ src/IO/S3/getObjectInfo.cpp
@@ -134,7 +134,7 @@ ObjectAttributes getObjectTags(
             error.GetErrorType(),
             "Failed to get object tags: {}. HTTP response code: {}.{}",
             error.GetMessage(),
-            static_cast<size_t>(error.GetResponseCode()),
+            error.GetResponseCode(),
             getAuthenticationErrorHint(error.GetErrorType()));
     }
 
@@ -165,7 +165,7 @@ ObjectInfo getObjectInfoIfExists(
         error.GetErrorType(),
         "Failed to get object info: {}. HTTP response code: {}.{}",
         error.GetMessage(),
-        static_cast<size_t>(error.GetResponseCode()),
+        error.GetResponseCode(),
         getAuthenticationErrorHint(error.GetErrorType()));
 }
 
@@ -188,7 +188,7 @@ ObjectInfo getObjectInfo(
         error.GetErrorType(),
         "Failed to get object info: {}. HTTP response code: {}.{}",
         error.GetMessage(),
-        static_cast<size_t>(error.GetResponseCode()),
+        error.GetResponseCode(),
         getAuthenticationErrorHint(error.GetErrorType()));
 }
 
@@ -219,7 +219,7 @@ bool objectExists(
 
     throw S3Exception(error.GetErrorType(),
         "Failed to check existence of key {} in bucket {}: {}. HTTP response code: {}, error type: {}.{}",
-        key, bucket, error.GetMessage(), static_cast<size_t>(error.GetResponseCode()),
+        key, bucket, error.GetMessage(), error.GetResponseCode(),
         error.GetErrorType(), getAuthenticationErrorHint(error.GetErrorType()));
 }
